+++ changelog.md
@@ -2,4 +2,7 @@
 
 ## v0.1.0
 
-* Add NotImplementError, NotExistedError, ExistedError
+* Add error types: NotImplemented, NotFound, Conflict, InfiniteLoop, Useless,
+  Capacity
+* Add data structures: Queue, Stack
+* Add other components: Deferral, Stopwatch